# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adam's Journal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PinInput {
  pin: string;
}

export interface AuthResult {
  success: boolean;
  message: string;
}

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  babyName?: string | null;
  /** @nullable */
  babyBirthDate?: string | null;
}

export type EventType = (typeof EventType)[keyof typeof EventType];

export const EventType = {
  feeding: "feeding",
  sleep: "sleep",
  diaper: "diaper",
} as const;

/**
 * @nullable
 */
export type EventDiaperType =
  | (typeof EventDiaperType)[keyof typeof EventDiaperType]
  | null;

export const EventDiaperType = {
  pee: "pee",
  poop: "poop",
  both: "both",
} as const;

export interface Event {
  id: number;
  type: EventType;
  startedAt: string;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  amountMl?: number | null;
  /** @nullable */
  diaperType?: EventDiaperType;
  /** @nullable */
  notes?: string | null;
  isActive: boolean;
}

export interface FeedingInput {
  /** @nullable */
  amountMl?: number | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
}

export interface SleepInput {
  startedAt: string;
  endedAt: string;
  /** @nullable */
  notes?: string | null;
}

export type DiaperInputDiaperType =
  (typeof DiaperInputDiaperType)[keyof typeof DiaperInputDiaperType];

export const DiaperInputDiaperType = {
  pee: "pee",
  poop: "poop",
  both: "both",
} as const;

export interface DiaperInput {
  diaperType: DiaperInputDiaperType;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  startedAt?: string | null;
}

/**
 * @nullable
 */
export type UpdateEventInputDiaperType =
  | (typeof UpdateEventInputDiaperType)[keyof typeof UpdateEventInputDiaperType]
  | null;

export const UpdateEventInputDiaperType = {
  pee: "pee",
  poop: "poop",
  both: "both",
} as const;

export interface UpdateEventInput {
  /** @nullable */
  amountMl?: number | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  diaperType?: UpdateEventInputDiaperType;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
}

export interface ActiveSleepResult {
  event?: Event;
  /** @nullable */
  elapsedMinutes?: number | null;
}

export interface DailySummary {
  date: string;
  feedingCount: number;
  totalFeedingMl: number;
  totalFeedingMinutes: number;
  sleepCount: number;
  totalSleepMinutes: number;
  diaperCount: number;
  peeCount: number;
  poopCount: number;
  feedingGoalMin: number;
  feedingGoalMax: number;
  sleepGoalMinutes: number;
}

export interface RecentActivity {
  /** @nullable */
  lastFeedingAt?: string | null;
  /** @nullable */
  lastSleepAt?: string | null;
  /** @nullable */
  lastDiaperAt?: string | null;
  /** @nullable */
  lastFeedingMinutesAgo?: number | null;
  /** @nullable */
  lastSleepMinutesAgo?: number | null;
  /** @nullable */
  lastDiaperMinutesAgo?: number | null;
}

export type ListEventsParams = {
  /**
   * ISO date string (YYYY-MM-DD), defaults to today
   */
  date?: string;
  limit?: number;
};

export type GetDailySummaryParams = {
  /**
   * ISO date string (YYYY-MM-DD), defaults to today
   */
  date?: string;
};