# lib/api-zod/src/generated/types/activeSleepResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adam's Journal API
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from "./event";

export interface ActiveSleepResult {
  event?: Event;
  /** @nullable */
  elapsedMinutes?: number | null;
}