# lib/api-zod/src/generated/types/authStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adam's Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  babyName?: string | null;
  /** @nullable */
  babyBirthDate?: string | null;
}