# lib/api-zod/src/generated/types/dailySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adam's Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface DailySummary {
  date: string;
  feedingCount: number;
  totalFeedingMl: number;
  totalFeedingMinutes: number;
  sleepCount: number;
  totalSleepMinutes: number;
  diaperCount: number;
  peeCount: number;
  poopCount: number;
  feedingGoalMin: number;
  feedingGoalMax: number;
  sleepGoalMinutes: number;
}