# lib/api-zod/src/generated/types/diaperInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adam's Journal API
 * OpenAPI spec version: 0.1.0
 */
import type { DiaperInputDiaperType } from "./diaperInputDiaperType";

export interface DiaperInput {
  diaperType: DiaperInputDiaperType;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  startedAt?: string | null;
}