# lib/api-zod/src/generated/types/sleepInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adam's Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface SleepInput {
  startedAt: string;
  endedAt: string;
  /** @nullable */
  notes?: string | null;
}